Move the education page to the industries section
