+++ app/messages/en.js
@@ -380,10 +380,6 @@ export default {
     title: 'A better way to collect customer payments',
     description: 'Join thousands of businesses who automate their Direct Debit collections with GoCardless.',
   },
-  education_dm: {
-    title: 'A simpler way for parents to pay',
-    description: 'Join the hundreds of schools and nurseries who use GoCardless to collect payments.',
-  },
   accountants: {
     title: 'GoCardless transforms cashflow for accountants',
     nav_title: 'Accountants',
@@ -398,6 +394,11 @@ export default {
     title: 'Direct Debit for Digital Agencies',
     nav_title: 'Digital Agencies',
     description: '',
+  },
+  education_dm: {
+    title: 'A simpler way for parents to pay',
+    nav_title: 'Education',
+    description: 'Join the hundreds of schools and nurseries who use GoCardless to collect payments.',
   },
   finance: {
     title: 'Direct Debit for Financial Services',
